requirements.txt
requests==2.10.0
pycrypto==2.6.1
Pillow==4.3.0
beautifulsoup4==4.6.0
mutagen==1.38
